+++ cvat/apps/engine/view_utils.py
@@ -20,7 +20,7 @@ from drf_spectacular.utils import extend_schema
 from cvat.apps.engine.mixins import UploadMixin
 from cvat.apps.engine.models import CloudStorage as CloudStorageModel
 from cvat.apps.engine.parsers import TusUploadParser
-from cvat.apps.iam.permissions import CloudStoragePermission
+from cvat.apps.engine.permissions import CloudStoragePermission
 
 
 def make_paginated_response(
